Clean code adjustments have been implemented as complained by coach
